Fix bug in DescribeRelease
